BREAKING: Refactor config handling
**Breaking Changes**

Use ARM64 runtime to match the default Powertools layer. To be
configurable in the future.

Switch to pydantic as it is more robust than my home brew solution. This
makes adding additional configuration options in the future easier.

Split the configuration loading from the config class. `ConfigLoader`
is responsible for loading the config. `Config` still managed the
configuration.

Refactored the tests to make them easier to maintain. Many are now
parameterised.

`vpc_id` is now required if using `subnets` in config. When using a shared
VPC the subnet data sources may not be available. While this adds some
redundant config it makes things more robust.